requirements.txt
python-dotenv
huggingface-hub
hf_transfer
